% Generated by roxygen2 (4.0.1): do not edit by hand
\name{crossAlert}
\alias{crossAlert}
\title{Alert for cross of two moving averages}
\usage{
crossAlert(f1, f2)
}
\arguments{
\item{f1}{Parameter for the slower exponential moving average}

\item{f2}{Parameter for the faster exponential moving average}
}
\value{
An unconditional TRUE value
}
\description{
This is am example function illustrating Streamulus
}
\author{
Dirk Eddelbuettel
}

