% Generated by roxygen2 (4.0.1): do not edit by hand
\name{helloStream}
\alias{helloStream}
\title{Print some hello messages}
\usage{
helloStream(ignored)
}
\arguments{
\item{ignored}{Just to supply a parameter}
}
\value{
An unconditional TRUE value
}
\description{
This is an example function illustrating Streamulus
}
\author{
Dirk Eddelbuettel
}

