% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{commonSubexpressions}
\alias{commonSubexpressions}
\title{Common 'greeting' carried across several expressions}
\usage{
commonSubexpressions()
}
\value{
An unconditional TRUE value
}
\description{
This is an example function illustrating Streamulus
}
\examples{
\dontrun{
commonSubexpressions()
}
}
\author{
Dirk Eddelbuettel
}
