% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{crossAlert}
\alias{crossAlert}
\title{Alert for cross of two moving averages}
\usage{
crossAlert(f1, f2)
}
\arguments{
\item{f1}{Parameter for the slower exponential moving average}

\item{f2}{Parameter for the faster exponential moving average}
}
\value{
An unconditional TRUE value
}
\description{
This is an example function illustrating Streamulus
}
\examples{
crossAlert(10,20)
}
\author{
Dirk Eddelbuettel
}

