% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{helloStream}
\alias{helloStream}
\title{Print some hello messages}
\usage{
helloStream()
}
\value{
An unconditional TRUE value
}
\description{
This is an example function illustrating Streamulus
}
\examples{
helloStream()
}
\author{
Dirk Eddelbuettel
}

