% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{slidingWindow}
\alias{slidingWindow}
\title{Sliding window function example}
\usage{
slidingWindow()
}
\value{
An unconditional TRUE value
}
\description{
This is an example function illustrating Streamulus
}
\examples{
\dontrun{
slidingWindow()
}
}
\author{
Dirk Eddelbuettel
}
